# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/user/": {
            "get": {
                "description": "allow people to view their user profile data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "View user data",
                "parameters": [
                    {
                        "type": "string",
                        "description": "page",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "limit",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            },
            "put": {
                "description": "allow people to update their user profile data",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "Update user data",
                "parameters": [
                    {
                        "description": "Add user",
                        "name": "new_user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.UsersModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            },
            "delete": {
                "description": "allow people to delete their account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Account"
                ],
                "summary": "url to delete user account",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/user/active_sessions/": {
            "get": {
                "description": "return the active user session/token across all browser",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "Get active user login session",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/user/block_token/": {
            "post": {
                "description": "Adds the token of user to block list based on provided token id",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "Block specified session",
                "parameters": [
                    {
                        "description": "block token",
                        "name": "block_active_session",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/database.ActiveSessionsModel"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
        "/user/logout/": {
            "get": {
                "description": "API allow user to logout, which delete the cookie which stores token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Session"
                ],
                "summary": "allow user to logout",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/my_modules.ResponseFormat"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "database.ActiveSessionsModel": {
            "type": "object",
            "properties": {
                "_id": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "exp": {
                    "type": "integer"
                },
                "ip": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "token_id": {
                    "type": "string"
                },
                "ua": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "database.UsersModel": {
            "type": "object",
            "required": [
                "access_level",
                "access_level_weight",
                "email",
                "mobile",
                "password",
                "username"
            ],
            "properties": {
                "access_level": {
                    "type": "string"
                },
                "access_level_weight": {
                    "type": "integer"
                },
                "auth_provider": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "mobile": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "uid": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
...
        "my_modules.ResponseFormat": {
            "type": "object",
            "required": [
                "data",
                "msg",
                "status"
            ],
            "properties": {
                "data": {},
                "msg": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    }
}